Add new PacketHandler and move changeStatus
