+++ app/src/main/java/com/geeksville/mesh/service/PacketHandler.kt
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2025 Meshtastic LLC
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.geeksville.mesh.service
+
+import com.geeksville.mesh.DataPacket
+import com.geeksville.mesh.MessageStatus
+import com.geeksville.mesh.concurrent.handledLaunch
+import com.geeksville.mesh.database.PacketRepository
+import dagger.Lazy
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.withTimeoutOrNull
+
+class PacketHandler(
+    private val packetRepository: Lazy<PacketRepository>,
+    private val serviceBroadcasts: MeshServiceBroadcasts,
+) {
+
+    private val scope = CoroutineScope(Dispatchers.IO)
+
+    /** Change the status on a DataPacket and update watchers */
+    fun changeStatus(packetId: Int, m: MessageStatus) = scope.handledLaunch {
+        if (packetId != 0) {
+            getDataPacketById(packetId)?.let { p ->
+                if (p.status == m) return@handledLaunch
+                packetRepository.get().updateMessageStatus(p, m)
+                serviceBroadcasts.broadcastMessageStatus(packetId, m)
+            }
+        }
+    }
+
+    private suspend fun getDataPacketById(packetId: Int): DataPacket? = withTimeoutOrNull(1000) {
+        var dataPacket: DataPacket? = null
+        while (dataPacket == null) {
+            dataPacket = packetRepository.get().getPacketById(packetId)?.data
+            if (dataPacket == null) delay(100)
+        }
+        dataPacket
+    }
+}
